[Android] Clean up MVI pattern
