+++ Android/app/src/main/java/com/coyotebitcoin/padawanwallet/presentation/viewmodels/LessonsViewModel.kt
@@ -5,18 +5,36 @@
 
 package com.coyotebitcoin.padawanwallet.presentation.viewmodels
 
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.coyotebitcoin.padawanwallet.domain.lessons.LessonsRepository
-import com.coyotebitcoin.padawanwallet.presentation.viewmodels.mvi.LessonAction
-import com.coyotebitcoin.padawanwallet.presentation.viewmodels.mvi.LessonsRootScreenState
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
 
 private const val TAG = "PTag/ChaptersViewModel"
+
+data class LessonsRootScreenState(
+    val completedLessons: Map<Int, Boolean> = mapOf(
+        Pair(1, false),
+        Pair(2, false),
+        Pair(3, false),
+        Pair(4, false),
+        Pair(5, false),
+        Pair(6, false),
+        Pair(7, false),
+        Pair(8, false),
+        Pair(9, false),
+    )
+)
+
+sealed interface LessonAction {
+    data class  SetCompleted(val lessonNum: Int) : LessonAction
+    data object ResetAll : LessonAction
+}
 
 class LessonsViewModel : ViewModel() {
     private val initialCompletedChaptersState: Map<Int, Boolean>
@@ -29,8 +47,8 @@ class LessonsViewModel : ViewModel() {
         initialCompletedChaptersState = completedChapters
     }
 
-    var rootState: LessonsRootScreenState by mutableStateOf(LessonsRootScreenState(initialCompletedChaptersState))
-        private set
+    private val _rootState = MutableStateFlow(LessonsRootScreenState(initialCompletedChaptersState))
+    val rootState: StateFlow<LessonsRootScreenState> = _rootState.asStateFlow()
 
     fun onAction(action: LessonAction) {
         when (action) {
@@ -45,13 +63,13 @@ class LessonsViewModel : ViewModel() {
             LessonsRepository.setCompletedLesson(chapterNum)
             completedChapters = LessonsRepository.getCompletedLessons()
         }
-        rootState = LessonsRootScreenState(completedChapters.toMap())
+        _rootState.update { it.copy(completedChapters.toMap()) }
     }
 
     private fun resetCompletedLessons() {
         viewModelScope.launch {
             LessonsRepository.resetCompletedLessons()
         }
-        rootState = LessonsRootScreenState()
+        _rootState.update { LessonsRootScreenState() }
     }
 }
